ALPR/requirements.txt
numpy
ffmpeg-python
opencv-python-headless
os
json
pytesseract
IMAGE
Pipeline
